rename "examples" to "demos" to avoid confusion with Cargo examples
I got confused because I tried to run the player with:

cargo run --examples demo_player

However this did not work because demo_player is structured as a
subcrate in a workspace, not as an example. Cargo examples do not
have a src directory nor their own Cargo.toml:
https://doc.rust-lang.org/cargo/guide/project-layout.html

Unfortunately there is no way to specify dependencies only for an
example. There is an open RFC for that:
https://github.com/rust-lang/rfcs/pull/2887

So for now, I think it is best to leave the demo applications as
subcrates in a workspace, but rename the "examples" directory to
"demos" so people don't think they will work with
cargo run --examples